refactor(contour): inline zoom parameter helper
